Elevator that extends TrapezoidProfileSubsystem
This is my first try at learning how to use the TrapezoidProfileSubsystem.

The base class handles the periodic method and has a setGoal(double goal) method that must be called to set the goal. This is done with the setElevatorPosition command.

The base class has a useState(State) method that must be overriden. This is where the updated TrapezoidProfile State is passed in. This is where the motor controllers are called to go to the target setpoint.

The ElevatorPosition enum encapsulates the fact that the elevator has exactly 2 positions.

The FeedforwardGains is an example of how to construct a complex data object using the Builder pattern. This pattern ensures that the data object is immutable. This pattern is preferred over a complex set of constructors where 0 or more parameters must be provided in a pre-defined order. Finally, this will fail any attempts to create data objects with negative gains.